Rename controller package to metrics
We have similar packages such as controller and controllers, which makes
us confusing. controller package is used by controller sub command,
therefore the name is as. But, it is actually related to metrics, so I
change its name to metrics. Also, runners package has too few logic, I
embed it into cmd/controller.go directly.
