+++ metrics/metrics.go
@@ -1,4 +1,4 @@
-package controller
+package metrics
 
 import (
 	"github.com/prometheus/client_golang/prometheus"
